+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [1.23.0] - 2020-12-09
+
+## Changed
+
+- serviceId added for configuring contact-frontend for consistency with
+contact-frontend documentation, serviceCode deprecated
+
+## Fixed
 
 - Fixed colour contrast for focused sign out link in gov.uk header not meeting accessibility standards